Other: Merge configuration defined converters into one file.
